[#117] feat - 서버 수정 기능 추가
230730 김수현

+++ src/main/java/com/ssafy/withview/controller/ServerController.java
@@ -21,7 +21,7 @@ import java.util.List;
 import java.util.UUID;
 
 @RestController
-@RequestMapping("/api/server")
+@RequestMapping("/api/servers")
 @RequiredArgsConstructor
 public class ServerController {
 	private final ServerService serverService;
@@ -152,36 +152,39 @@ public class ServerController {
 		System.out.println("====== 서버 추가 끝 ======");
 		return new ResponseEntity<JSONObject>(jsonObject, HttpStatus.OK);
 	}
-	@PutMapping("/")
+	@PutMapping("")
 	public ResponseEntity<?> updateServer(@ModelAttribute ServerDto serverDto, @RequestParam(name = "file", required = false) MultipartFile multipartFile) {
 		System.out.println("====== 서버 추가 시작 ======");
 		JSONObject jsonObject = new JSONObject();
 		try{
-
 			// #1 - 버킷 생성
-			if (!s3client.doesBucketExist(bucketName)) {
-				s3client.createBucket(bucketName);
+			if(multipartFile != null){
+				System.out.println("=== 파일 변경 ===");
+				if (!s3client.doesBucketExist(bucketName)) {
+					s3client.createBucket(bucketName);
+				}
+				// #2 - 원본 파일 이름 저장
+				String originalName = multipartFile.getOriginalFilename();
+				serverDto.setBackgroundImgOriginalName(originalName);
+
+				// #3 - 저장용 랜점 파일 이름 저장
+				String extend = originalName.substring(originalName.lastIndexOf('.'));
+				UUID uuid = UUID.randomUUID();
+				String backgroundImgSearchName = uuid.toString()+extend;
+
+				// #4 - 파일 임시 저장
+				File backgroundImgFile = new File(resourceLoader.getResource("classpath:/img/").getFile().getAbsolutePath(),backgroundImgSearchName);
+				multipartFile.transferTo(backgroundImgFile);
+
+				// #5 - 이미지 서버 저장
+				s3client.putObject(bucketName, "server-background/"+backgroundImgSearchName, backgroundImgFile);
+				jsonObject.put("imgUrl",CLOUD_FRONT_URL+"server-background/"+backgroundImgSearchName);
+				// #6 - DB 저장
+				serverDto.setBackgroundImgSearchName(uuid.toString()+extend);
+				backgroundImgFile.delete();	//기존 임시 저장용 파일 삭제
 			}
-			// #2 - 원본 파일 이름 저장
-			String originalName = multipartFile.getOriginalFilename();
-			serverDto.setBackgroundImgOriginalName(originalName);
-
-			// #3 - 저장용 랜점 파일 이름 저장
-			String extend = originalName.substring(originalName.lastIndexOf('.'));
-			UUID uuid = UUID.randomUUID();
-			String backgroundImgSearchName = uuid.toString()+extend;
-
-			// #4 - 파일 임시 저장
-			File backgroundImgFile = new File(resourceLoader.getResource("classpath:/img/").getFile().getAbsolutePath(),backgroundImgSearchName);
-			multipartFile.transferTo(backgroundImgFile);
-
-			// #5 - 이미지 서버 저장
-			s3client.putObject(bucketName, "server-background/"+backgroundImgSearchName, backgroundImgFile);
-			jsonObject.put("imgUrl",CLOUD_FRONT_URL+"server-background/"+backgroundImgSearchName);
-			// #6 - DB 저장
-			serverDto.setBackgroundImgSearchName(uuid.toString()+extend);
 			serverDto = serverService.updateServer(serverDto);
-			backgroundImgFile.delete();	//기존 임시 저장용 파일 삭제
+
 		}catch (Exception e){
 			e.printStackTrace();
 			jsonObject.put("success",false);
